chore: relocate crypto utils
