+++ packages/auth/src/index.ts
@@ -1,2 +1,3 @@
 export { Auth, IAuth, IAuthWebUI } from './auth';
 export * from './utils';
+export * from './crypto-utils';
