+++ packages/utils/package.json
@@ -19,7 +19,6 @@
     "@verdaccio/dev-commons": "workspace:5.0.0-alpha.0",
     "@verdaccio/readme": "workspace:*",
     "js-yaml": "3.13.1",
-    "jsonwebtoken": "8.5.1",
     "minimatch": "3.0.4",
     "semver": "7.3.2"
   },
