+++ packages/utils/src/crypto-utils.ts
@@ -1,35 +1,8 @@
-import { createDecipher, createCipher, createHash, pseudoRandomBytes, Hash } from 'crypto';
-import jwt from 'jsonwebtoken';
+import { createHash, pseudoRandomBytes, Hash } from 'crypto';
 
-import { JWTSignOptions, RemoteUser } from '@verdaccio/types';
-
-export const defaultAlgorithm = 'aes192';
 export const defaultTarballHashAlgorithm = 'sha1';
 
-export function aesEncrypt(buf: Buffer, secret: string): Buffer {
-  // deprecated (it will be migrated in Verdaccio 5), it is a breaking change
-  // https://nodejs.org/api/crypto.html#crypto_crypto_createcipher_algorithm_password_options
-  // https://www.grainger.xyz/changing-from-cipher-to-cipheriv/
-  const c = createCipher(defaultAlgorithm, secret);
-  const b1 = c.update(buf);
-  const b2 = c.final();
-  return Buffer.concat([b1, b2]);
-}
-
-export function aesDecrypt(buf: Buffer, secret: string): Buffer {
-  try {
-    // deprecated (it will be migrated in Verdaccio 5), it is a breaking change
-    // https://nodejs.org/api/crypto.html#crypto_crypto_createdecipher_algorithm_password_options
-    // https://www.grainger.xyz/changing-from-cipher-to-cipheriv/
-    const c = createDecipher(defaultAlgorithm, secret);
-    const b1 = c.update(buf);
-    const b2 = c.final();
-    return Buffer.concat([b1, b2]);
-  } catch (_) {
-    return new Buffer(0);
-  }
-}
-
+// podria moverse a storage donde se usa
 export function createTarballHash(): Hash {
   return createHash(defaultTarballHashAlgorithm);
 }
@@ -41,40 +14,12 @@ export function createTarballHash(): Hash {
  * @param {Object} data
  * @return {String}
  */
+// se usa en api, middleware, web
 export function stringToMD5(data: Buffer | string): string {
   return createHash('md5').update(data).digest('hex');
 }
 
+// se usa en config
 export function generateRandomHexString(length = 8): string {
   return pseudoRandomBytes(length).toString('hex');
-}
-
-/**
- * Sign the payload and return JWT
- * https://github.com/auth0/node-jsonwebtoken#jwtsignpayload-secretorprivatekey-options-callback
- * @param payload
- * @param secretOrPrivateKey
- * @param options
- */
-export async function signPayload(
-  payload: RemoteUser,
-  secretOrPrivateKey: string,
-  options: JWTSignOptions = {}
-): Promise<string> {
-  return new Promise(function (resolve, reject): Promise<string> {
-    return jwt.sign(
-      payload,
-      secretOrPrivateKey,
-      {
-        // 1 === 1ms (one millisecond)
-        notBefore: '1', // Make sure the time will not rollback :)
-        ...options,
-      },
-      (error, token) => (error ? reject(error) : resolve(token))
-    );
-  });
-}
-
-export function verifyPayload(token: string, secretOrPrivateKey: string): RemoteUser {
-  return jwt.verify(token, secretOrPrivateKey);
 }
